bugfix: QFree lowering (#167)
Guppy `discard` -> HUGR `QFree` was being lowered to
`__quantum__rt__qubit_release`, which is not supported on h-series.

Changes:
- With `target="quantinuum-hardware"` emit nothing for `QFree`. For
`target="native"`, continue to emit `__quantum__rt__qubit_release`
- As a bit of a hack, `qircheck` errors with error message containing
`__quantum__rt__qubit_release` are disregarded if `target="native"`,
otherwise `qircheck` would fail. Not sure if we actually want this or if
we should stipulate that `qircheck` only works for
`target="quantinuum-hardware"`...
- Added a test guppy example with `discard`